Rearrange some functions and methods
