+++ internal/s3storage/artifactreadstorage.go
@@ -2,17 +2,13 @@ package s3storage
 
 import (
 	"context"
-	"io/ioutil"
-	"os"
 	"path"
 	"sort"
 
 	"github.com/aws/aws-sdk-go/aws"
 	"github.com/aws/aws-sdk-go/aws/awserr"
 	"github.com/aws/aws-sdk-go/service/s3"
-	"github.com/aws/aws-sdk-go/service/s3/s3manager"
 	"github.com/lunarway/release-manager/internal/artifact"
-	"github.com/lunarway/release-manager/internal/log"
 	"github.com/pkg/errors"
 )
 
@@ -95,24 +91,6 @@ func (f *Service) ArtifactSpecifications(ctx context.Context, service string, n 
 	return artifactSpecs, nil
 }
 
-func (f *Service) getLatestObjectKey(ctx context.Context, service string, branch string) (string, error) {
-	list, err := f.s3client.ListObjectsV2(&s3.ListObjectsV2Input{
-		Bucket:  aws.String(f.bucketName),
-		MaxKeys: aws.Int64(1000), // TODO: Find a solution to handle more than 1000
-		Prefix:  aws.String(getServiceAndBranchObjectKeyPrefix(service, branch)),
-	})
-
-	if err != nil {
-		return "", err
-	}
-
-	sort.Slice(list.Contents, func(i, j int) bool {
-		return list.Contents[i].LastModified.After(*list.Contents[j].LastModified)
-	})
-
-	return *list.Contents[0].Key, nil
-}
-
 func (f *Service) getArtifactSpecFromObjectKey(ctx context.Context, objectKey string) (artifact.Spec, error) {
 	head, err := f.s3client.HeadObject(&s3.HeadObjectInput{
 		Bucket: aws.String(f.bucketName),
@@ -129,68 +107,4 @@ func (f *Service) getArtifactSpecFromObjectKey(ctx context.Context, objectKey st
 	}
 
 	return artifactSpec, nil
-}
-
-// downloadArtifact downloads an artifact from its AWS S3 key. The returned
-// string is a file system path to a raw artifact, ie. unzipped.
-func (f *Service) downloadArtifact(ctx context.Context, key string) (string, func(context.Context), error) {
-	logger := log.WithContext(ctx)
-
-	zipDestPath, closeSource, err := tempDir("s3-artifact-paths")
-	if err != nil {
-		return "", nil, errors.WithMessage(err, "get temp dir")
-	}
-	defer closeSource(ctx)
-	zipDestPath = path.Join(zipDestPath, "artifact.zip")
-	logger.Debugf("Zip destination: %s", zipDestPath)
-	zipDest, err := os.Create(zipDestPath)
-	if err != nil {
-		return "", nil, errors.WithMessage(err, "get temp dir")
-	}
-	defer func() {
-		err := zipDest.Close()
-		if err != nil {
-			logger.Errorf("Failed to close zip destination file: %v", err)
-		}
-	}()
-
-	downloader := s3manager.NewDownloaderWithClient(f.s3client)
-	logger.Infof("Downloading object at key '%s'", key)
-	n, err := downloader.DownloadWithContext(ctx, zipDest, &s3.GetObjectInput{
-		Bucket: aws.String(f.bucketName),
-		Key:    aws.String(key),
-	})
-	if err != nil {
-		return "", nil, errors.WithMessage(err, "download object")
-	}
-	logger.Infof("Downloaded %d bytes", n)
-
-	destPath, closeSource, err := tempDir("s3-artifact-paths")
-	if err != nil {
-		return "", nil, errors.WithMessage(err, "get temp dir")
-	}
-	logger.Infof("Resources dest: %s", destPath)
-
-	files, err := unzipFile(zipDestPath, destPath)
-	if err != nil {
-		// manually close destination directory here as we must allow callers to
-		// access the directory and this cannot use defer
-		closeSource(ctx)
-		return "", nil, errors.WithMessage(err, "unzip file")
-	}
-	logger.Infof("Found files: %v", files)
-	return destPath, closeSource, nil
-}
-
-func tempDir(prefix string) (string, func(context.Context), error) {
-	path, err := ioutil.TempDir("", prefix)
-	if err != nil {
-		return "", func(context.Context) {}, err
-	}
-	return path, func(ctx context.Context) {
-		err := os.RemoveAll(path)
-		if err != nil {
-			log.WithContext(ctx).Errorf("Removing temporary directory failed: path '%s': %v", path, err)
-		}
-	}, nil
 }
